server.ts now creates a single DocumentSymbolStore and registers the three providers via: registerCompletionProvider(...) registerHoverProvider(...) registerFormattingProvider(...)
