+++ server/tsconfig.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/documentation.ts","./src/formatter.ts","./src/parsepipeline.ts","./src/prototypes.ts","./src/reservedwords.ts","./src/server.ts","./src/symbols.ts","./src/typedocumentation.ts","./src/validator.ts","./src/antlr/proglang12dlexer.ts","./src/antlr/proglang12dlistener.ts","./src/antlr/proglang12dparser.ts","./src/antlr/proglang12dvisitor.ts"],"errors":true,"version":"5.9.3"}
+{"root":["./src/documentation.ts","./src/formatter.ts","./src/includes.ts","./src/parsepipeline.ts","./src/prototypes.ts","./src/reservedwords.ts","./src/server.ts","./src/symbols.ts","./src/typedocumentation.ts","./src/validator.ts","./src/antlr/proglang12dlexer.ts","./src/antlr/proglang12dlistener.ts","./src/antlr/proglang12dparser.ts","./src/antlr/proglang12dvisitor.ts","./src/providers/completionprovider.ts","./src/providers/documentsymbols.ts","./src/providers/formattingprovider.ts","./src/providers/hoverprovider.ts","./src/providers/utils.ts"],"errors":true,"version":"5.9.3"}
